### 2023-08-29 #604 @HaoZeke
cuh2vizR: Visualizing and reasoning about ND potential energy surfaces
Submitting Author Name: Rohit Goswami
Submitting Author Github Handle: @HaoZeke
Repository:  https://github.com/HaoZeke/cuh2vizR
Submission type: Pre-submission

---

```
Package: cuh2vizR
Title: Plot Copper-Hydrogen Systems
Version: 0.0.0.9000
    person("Rohit", "Goswami", , "rgoswami@ieee.org", role = c("aut", "cre"),
Description: A package meant to plot a CuH2 system
SystemRequirements: C++17
    cpp11
    assertthat,
    cli,
    gganimate,
    glue,
    optparse,
    progress,
    readConR
```

So this is a wrapper around a potential energy function written in Fortran, wrapped in C++ and then passed through to R. It can be used to read `con` files (though there's a separate package I'm working on for that). The main user benefit is the ability to quickly plot the energy surface of a non-trivial chemical system, that of Hydrogen molecule diffusion on a copper surface. This is an ND system, but with only two degrees of freedom, making it good to visualize.

Downstream users (my group, collaborators) use this to train / view and reason about potential energy surface surrogates. It make it easy to visualize a normally complex system and compute metrics. Since the eyeball norm is popular, and often we train over multiple system points, this also provides a CLI with animation capability for on-the-fly visuals.

To the best of my knowledege there are no R packages tailored to materials science / physical chemistry use. There are python packages (e.g. ase, the atomic simulation environment), but they are far too slow. Indeed, speed is one of the main considerations of [this package](https://github.com/HaoZeke/cuh2vizR/issues/3).

I'd love to get reviews and there's definitely a lot of science going on (potential energy functions are hard!) but the libraries make the R wrapper unlikely to make it to CRAN (there are quite a few dependencies and its best run with `micromamba` or `pixi`). I've tested it (and my collaborators) on several systems (MacOS and *nix systems including HPCs), but that was why I'm not sure it'd be a good fit.

I know there's some more work to be done (vignettes), but I was hoping to get a go/no-go for submission via the presubmission system :)
