### 2023-08-29 #604 @noamross
Hello @HaoZeke, thank you for your submission!  This is an interesting case. We are certainly interested in supporting tools of materials science and physical chemistry tools.

First, while we don't require that things will go to CRAN, we have similar requirements that the package should pass R CMD CHECK tests on CI across platforms.  That this would require setting organizing the package to usuable "standalone", without the pixi or micromamba, following R package conventions for including compiled code.  As far as I can tell from browsing around your repositories, this is is doable with some reorganization but involve different build tooling that you have been using.

Secondly, I note that there are not a lot of functions in this package, given that the bulk of the development action is in `readCon` and `potlib`, and they are mostly for visualization.  Straight visualization tools are out of scope for us unless they are part of the [statistical EDA](https://stats-devguide.ropensci.org/standards.html#standards-eda). The `readCon` package is a more straightforward fit, in that it is for "parsing scientific data types and outputs from scientific equipment", under our **data extraction category**.  Looking at both, I think it would make sense to fold these R packages together.

Finally, I see that you've put a good bit of the development effort into the CLI, rather than R interface, for these functions.  There isn't a lot of convention in this area, so I'd anticipate a good bit of discussion from reviewers about the best way to organize this.  I'd encourage you to look at how it is done in these examples: https://github.com/eddelbuettel/littler/tree/master/inst/examples

So in summary - `cuh2vizR` is an edge case, but `readConR` or a combination of the two fits into our scope, and we'd be happy to review it if it can meet our testing and CI requirements.
