### 2023-08-29 #604 @HaoZeke
Thank you for the detailed comments (and prompt response!). I will address each in order

> Hello @HaoZeke, thank you for your submission! This is an interesting case. We are certainly interested in supporting tools of materials science and physical chemistry tools.
>
> First, while we don't require that things will go to CRAN, we have similar requirements that the package should pass R CMD CHECK tests on CI across platforms. That this would require setting organizing the package to usuable "standalone", without the pixi or micromamba, following R package conventions for including compiled code. As far as I can tell from browsing around your repositories, this is is doable with some reorganization but involve different build tooling that you have been using.

I think that makes sense, especially `readCon`, will need to think a bit about `potlib`. `pixi` and `micromamba` are essentially helpers, without which the user must have the required libraries system wide.

> Secondly, I note that there are not a lot of functions in this package, given that the bulk of the development action is in `readCon` and `potlib`, and they are mostly for visualization. Straight visualization tools are out of scope for us unless they are part of the [statistical EDA](https://stats-devguide.ropensci.org/standards.html#standards-eda). The `readCon` package is a more straightforward fit, in that it is for "parsing scientific data types and outputs from scientific equipment", under our **data extraction category**. Looking at both, I think it would make sense to fold these R packages together.

Unfortunately I don't think it would be a good idea to fold cuh2vizR into either `readCon` or `potlib`, since the bulk of the functionality is geared towards a very specific chemical system (a very useful test case, but nonetheless not general enough to be in either). I'm glad to hear the `readCon` package would make sense!

> Finally, I see that you've put a good bit of the development effort into the CLI, rather than R interface, for these functions. There isn't a lot of convention in this area, so I'd anticipate a good bit of discussion from reviewers about the best way to organize this. I'd encourage you to look at how it is done in these examples: https://github.com/eddelbuettel/littler/tree/master/inst/examples
>
> So in summary - `cuh2vizR` is an edge case, but `readConR` or a combination of the two fits into our scope, and we'd be happy to review it if it can meet our testing and CI requirements.

I think then, I shall clean up `readconR` ([here](https://github.com/HaoZeke/readConR)) for submission first, and then try to see if I can get `potlibR` to compile in isolation for a possible followup.

However, one tiny note, the main functionality of `cuh2vizR` is one of data-wrangling, given a `readCon` output data-frame, this generates one point of an energy surface, to get the other points, `cuh2vizR` has helpers to convert the user friendly (and intuitive) concepts of H-H distance and Cu-H distance to the number of atoms * 3 matrix format and has helpers to generate a grid over H-H and Cu-H values to study the entire energy surface (formed by evaluating for each point the system configuration).

This is also why it is specialized to a single system. Every potential in `potlib` will provide an energy for a configuration read with `readcon`, but *only* for the CuH2 case does it make sense to plot in 2 dimensions because the effective degrees of freedom in the system are two, and so only for such a system does one need the remaining helpers (which are not exposed to the user, in R except through the generation of a data-frame), but could be. Does this sound like it might be suitable? If not I'll close this and reopen with `readcon` :)
