### 2023-08-30 #604 @noamross
Thanks for the explanation. In general we aim to balance generality for re-use with field-specific tooling. I admit I have a bit of trouble with physical chemistry being far my field, but based on your explanation, `readCon` provides a scientific format parser (👍 in scope), `potlib` provides a set of standard and expandable library potential calculations/values (👍 in-scope), and `cuh2vizR` provides a visualization for a narrow case where the dimension is 2.  The thing I don't fully understand is how much `cuh2vizR` is a general tool vs a narrow workflow.

One reason data visualization tends to be out-of-scope is it tends to be difficult to standardize - people have various and strong opinions as to what plots should look like, and it is often quite specific to the application and venue.   So we often encourage authors to produce processed data in a standard format ready for visualization by the user, and/or plot objects designed to be modified  It seems to me that a standard framework that works for higher dimensions, with visualization options for 2-D cases or 2-D slices, would be best, but that's really for input from reviewers we find with some experience in your field.

But yes, I'd suggest preparing `readConR` for submission, and deferring `cuh2vizR`, which might make more sense folded in to `potlibR`.
