### 2023-08-31 #604 @noamross
Yes, that would be in scope, under the "data munging" category if the manipulations are sufficiently standard calculations.  If the manipulations are novel algorithms, it's the kind of thing that would fall under our [statistical standards](https://stats-devguide.ropensci.org/standards.html) though admittedly those may not be a great match for some things in the physical sciences.
